fix: typo in the directory name of default-theme
