+++ components/resume/themes/DEFAULT_THEME/sections/EducationTitle.tsx
@@ -1,4 +1,4 @@
-import {Divider, SectionTitle} from '@/components/resume/themes/DEAFULT_THEME/shared/Components'
+import {Divider, SectionTitle} from '@/components/resume/themes/DEFAULT_THEME/shared/Components'
 
 const EducationTitle = () => {
 	return (
